Address review comments: pin RPM runner, exclude docs from triggers, inline wheel helpers, and register manylinux workflow as skippable CI
